// backend/shared/models/UserInteraction.d.ts
import mongoose, { Schema } from 'mongoose';
import { IUserInteraction } from '../types';
export declare const UserInteraction: mongoose.Model<IUserInteraction, {}, {}, {}, mongoose.Document<unknown, {}, IUserInteraction> & IUserInteraction & Required<{
    _id: Schema.Types.ObjectId;
}>, any>;
//# sourceMappingURL=UserInteraction.d.ts.map